[feat] Better redux actions structure
